align with java example
