+++ documentation-samples/quickstarts/get-answer/QnAMakerAnswerQuestion/QnAMakerAnswerQuestion.csproj
